remove fallback from {{tl|YoutubeCount}}; change image cropping algorithm

+++ process_image.py
@@ -8,7 +8,7 @@ def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("-i", dest="input", type=Path)
     parser.add_argument("-o", dest="output", type=Path)
-    parser.add_argument("-t", dest="threshold", type=int)
+    parser.add_argument("-t", dest="threshold", type=float)
     args = parser.parse_args()
     file_in: Path = args.input
     file_out: Path = args.output
